Invoice reconciliation App/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-openai
openai
pdfplumber
requests
streamlit
python-dotenv
langchain_community
python-multipart

psycopg2-binary 
sqlalchemy 
scikit-learn
streamlit-ace